+++ client/src/components/admin/dashboard/Dashboard.js
@@ -62,8 +62,9 @@ const Dashboard = ({
             </h6>
 
             <div className="d-flex w-100 align-items-center justify-content-end ">
+            
               {/* Year Filter */}
-              <div className="p-0 mx-2" style={{ maxWidth: "200px" }}>
+              {/* <div className="p-0 mx-2" style={{ maxWidth: "200px" }}>
                 <select
                   className="form-select form-control bg-primary text-white m-0"
                   aria-label="Default select example"
@@ -78,10 +79,10 @@ const Dashboard = ({
                   <option value={2023}>2022-2023</option>
                   <option value={2024}>2023-2024</option>
                 </select>
-              </div>
+              </div> */}
 
               {/* Grade Filter */}
-              <div className="p-0 mx-2" style={{ maxWidth: "200px" }}>
+              {/* <div className="p-0 mx-2" style={{ maxWidth: "200px" }}>
                 <select
                   className="form-select form-control bg-primary text-white m-0"
                   aria-label="Default select example"
@@ -97,10 +98,10 @@ const Dashboard = ({
                   <option value={5}>Room 505</option>
                   <option value={6}>Room 506</option>
                 </select>
-              </div>
+              </div> */}
 
               {/* Benchmark Filter */}
-              <div className="p-0" style={{ maxWidth: "200px" }}>
+              {/* <div className="p-0" style={{ maxWidth: "200px" }}>
                 <select
                   className="form-select form-control bg-primary text-white m-0"
                   aria-label="Default select example"
@@ -114,10 +115,10 @@ const Dashboard = ({
                   <option value={3}>Trivias</option>
                   <option value={4}>Jokes</option>
                 </select>
-              </div>
+              </div> */}
 
               {/* Benchmark Filter */}
-              <div className="p-0 mx-2" style={{ maxWidth: "200px" }}>
+              {/* <div className="p-0 mx-2" style={{ maxWidth: "200px" }}>
                 <select
                   className="form-select form-control bg-primary text-white m-0"
                   aria-label="Default select example"
@@ -131,29 +132,29 @@ const Dashboard = ({
                   <option value={3}>Aden 4</option>
                   <option value={4}>Joshua 1</option>
                 </select>
-              </div>
+              </div> */}
 
               {/* Report */}
-              <div className="p-0 mx-2" style={{ maxWidth: "200px" }}>
+              {/* <div className="p-0 mx-2" style={{ maxWidth: "200px" }}>
                 <DashboardExport reportName={`Statistics-${new Date().toLocaleString()}`}  reports={formatReport(statisticlogsList)} />
-              </div>
+              </div> */}
             </div>
           </div>
-          {reports !== undefined && (
+          {statisticlogsList !== undefined && (
             <div className="d-flex justify-content-between align-items-center flex-column flex-lg-row row">
               <div className="col-12">
                 <div className="row">
-                  <div className="col-4">
-                    <ChartPlaceHolder title={"Top Played Songs"} />
+                  <div className="col-12 col-lg-2">
+                    <ChartPlaceHolder title={"Statistics Log Report"} reportName={`Statistics-${new Date().toLocaleString()}`}  reports={formatReport(statisticlogsList)} />
                   </div>
 
-                  <div className="col-8">
+                  {/* <div className="col-8">
                     <ChartPlaceHolder title={"Top Read Books"} />
-                  </div>
+                  </div> */}
                 </div>
-                <div className="col-12">
+                {/* <div className="col-12">
                   <ChartPlaceHolder title={"Average Time Spending per Activity"} />
-                </div>
+                </div> */}
               </div>
             </div>
           )}